Detalles finales

+++ index.js
@@ -7,36 +7,50 @@ const app = express()
 const port = process.env.PORT || 3001
 
 app.set('llave', config.llave);
+app.use(bodyParser.urlencoded({extended: true}))
 app.use(bodyParser.json())
-app.use(bodyParser.urlencoded({extended: false}))
 
 const users = require("./user.js");
 
-/*app.post('/api/autenticar', (req, res) => {
-    return users.getByNickName(req.body.nickName)
+app.post('/api/user/add', function (request, response) {
+
+    return users.addUser(request)
+    .then(function (users) {
+        if (users) {
+            response.send(users);
+        } else {
+            response.status(400).send('Error in insert new record');
+        }
+    })
+    .catch(err=>response.status(400).send(err.parent.sqlMessage));        
+});
+
+app.post('/api/autenticar', (request, response) => {
+    return users.getByNickName(request.body.user)
     .then(function (users) {
         if (users){
-            if (users.toJSON().password==req.body.password){
+            if (users.toJSON().password==request.body.password){
                 const payload = {
-                    check:  true
-                   };
-                   const token = jwt.sign(payload, app.get('llave'), {
-                    expiresIn: 1440
-                   });
-                   res.json({
-                    mensaje: 'Autenticación correcta',
-                    token: token
-                   });
-            }else{
-                res.status(400).send('La contraseña que ha ingresado es incorrecta');
+                check:  true
+                };
+                const token = jwt.sign(payload, app.get('llave'), {
+                expiresIn: 1440
+                });
+                response.json({
+                mensaje: 'Autenticación correcta',
+                token: token
+                });
+            } else {
+                response.json({ mensaje: "Contraseña incorrectos"})
             }
         }else{
-            res.status(400).send('No existe el usuario ' + req.body.nickName);
+            response.json({ mensaje: "Usuario inexistente"})
         }
-    });
+    })
 })
 
-app.use((req, res, next) => {
+const rutasProtegidas = express.Router();
+rutasProtegidas.use((req, res, next) => {
     const token = req.headers['access-token'];
  
     if (token) {
@@ -53,12 +67,13 @@ app.use((req, res, next) => {
           mensaje: 'Token no proveída.' 
       });
     }
- });*/
+ });
+
 
 //Endpoints para Roles
 const roles = require("./role.js");
 
-app.get('/api/role/getAll', (request, response) => {
+app.get('/api/role/getAll', rutasProtegidas, (request, response) => {
     return roles.getAllRoles()
     .then(function (roles) {
         if (roles) {
@@ -69,7 +84,7 @@ app.get('/api/role/getAll', (request, response) => {
         }
     });
 })
-app.get('/api/role/getById/:id', (request, response)=>{
+app.get('/api/role/getById/:id', rutasProtegidas, (request, response)=>{
     return roles.getById(request.params.id)
     .then(function (roles) {
         if (roles) {
@@ -82,22 +97,8 @@ app.get('/api/role/getById/:id', (request, response)=>{
 
 
 //Endpoints para Users
-app.post('/api/user/add', function (request, response) {
 
-    return users.addUser(request)
-    .then(function (users) {
-        if (users) {
-            response.send(users);
-        } else {
-            response.status(400).send('Error in insert new record');
-        }
-    })
-    .catch(err=>response.status(400).send(err.parent.sqlMessage));
-    
-    
-        
-});
-app.put('/api/user/update/:id', function (request, response) {
+app.put('/api/user/update/:id', rutasProtegidas, function (request, response) {
     return users.updateUser(request)
     .then(function (users) {
         if (users) {
@@ -107,7 +108,7 @@ app.put('/api/user/update/:id', function (request, response) {
         }
     });    
 });
-app.delete('/api/user/delete/:id', function (request, response) {
+app.delete('/api/user/delete/:id', rutasProtegidas, function (request, response) {
     return users.deleteUser(request)
     .then(function (users) {
         if (users) {
@@ -117,7 +118,7 @@ app.delete('/api/user/delete/:id', function (request, response) {
         }
     });    
 });
-app.get('/api/user/getAll', (request, response) => {
+app.get('/api/user/getAll', rutasProtegidas, (request, response) => {
     return users.getAllUsers()
     .then(function (users) {
         if (users) {
@@ -128,7 +129,7 @@ app.get('/api/user/getAll', (request, response) => {
         }
     });
 })
-app.get('/api/user/getById/:id', (request, response)=>{
+app.get('/api/user/getById/:id', rutasProtegidas, (request, response)=>{
     return users.getById(request)
     .then(function (users) {
         if (users) {
@@ -138,13 +139,13 @@ app.get('/api/user/getById/:id', (request, response)=>{
         }
     });
 })
-app.get('/api/user/getByNickName/:nickName', (request, response)=>{
-    return users.getByNickName(request)
+app.get('/api/user/getByNickName', rutasProtegidas, (request, response)=>{
+    return users.getByNickName(request.body.nickName)
     .then(function (users) {
         if (users) {
             response.send(users);
         } else {
-            response.status(400).send('Error');
+            response.status(400).send('No existe el nickName ingresado');
         }
     });
 })
@@ -152,7 +153,7 @@ app.get('/api/user/getByNickName/:nickName', (request, response)=>{
 //Endpoints para Dishes
 const dishes = require("./dish.js");
 
-app.post('/api/dish/add', function (request, response) {
+app.post('/api/dish/add', rutasProtegidas, function (request, response) {
     return users.getByNickName(request.body.requestedBy)
     .then(function (users) {
         if (users){
@@ -172,32 +173,58 @@ app.post('/api/dish/add', function (request, response) {
         }else{
             response.status(400).send('No existe el usuario ' + request.body.requestedBy);
         }
-    });
-        
+    });       
         
 });
 
-app.put('/api/dish/update/:id', function (request, response) {
-    return dishes.updateDish(request)
-    .then(function (dishes) {
-        if (dishes) {
-            response.status(200).send('El plato ' + request.params.id + ' ha sido actualizado con éxito')
-        } else {
-            response.status(400).send('Error in update new record');
+app.put('/api/dish/update/:id', rutasProtegidas, function (request, response) {
+    return users.getByNickName(request.body.requestedBy)
+    .then(function (users) {
+        if (users){
+            if (users.toJSON().role==1){
+                return dishes.updateDish(request)
+                .then(function (dishes) {
+                    if (dishes) {
+                        response.status(200).send('El plato ' + request.params.id + ' ha sido actualizado con éxito')
+                    } else {
+                        response.status(400).send('Error in update new record');
+                    }
+                })    
+                .catch(err=>response.status(400).send(err.parent.sqlMessage));  
+            }else{
+                response.status(400).send('El usuario ' + request.body.requestedBy + ' no posee rol de administrador');
+            }
+        }else{
+            response.status(400).send('No existe el usuario ' + request.body.requestedBy);
         }
-    });    
+    });       
+        
 });
-app.delete('/api/dish/delete/:id', function (request, response) {
-    return dishes.deleteDish(request)
-    .then(function (dishes) {
-        if (dishes) {
-            response.status(200).send('El plato ' + request.params.id + ' ha sido eliminado con éxito')
-        } else {
-            response.status(400).send('Error in delete new record');
+
+app.delete('/api/dish/delete/:id', rutasProtegidas, function (request, response) {
+    return users.getByNickName(request.body.requestedBy)
+    .then(function (users) {
+        if (users){
+            if (users.toJSON().role==1){
+                return dishes.deleteDish(request)
+                .then(function (dishes) {
+                    if (dishes) {
+                        response.status(200).send('El plato ' + request.params.id + ' ha sido eliminado con éxito')
+                    } else {
+                        response.status(400).send('Error in delete new record');
+                    }
+                })    
+                .catch(err=>response.status(400).send(err.parent.sqlMessage));  
+            }else{
+                response.status(400).send('El usuario ' + request.body.requestedBy + ' no posee rol de administrador');
+            }
+        }else{
+            response.status(400).send('No existe el usuario ' + request.body.requestedBy);
         }
-    });    
+    });       
+        
 });
-app.get('/api/dishes/getAll', (request, response) => {
+app.get('/api/dish/getAll', rutasProtegidas, (request, response) => {
     return dishes.getAllDishes()
     .then(function (dishes) {
         if (dishes) {
@@ -208,7 +235,7 @@ app.get('/api/dishes/getAll', (request, response) => {
         }
     });
 })
-app.get('/api/dish/getById/:id', (request, response)=>{
+app.get('/api/dish/getById/:id', rutasProtegidas, (request, response)=>{
     return dishes.getById(request)
     .then(function (dishes) {
         if (dishes) {
@@ -223,30 +250,19 @@ app.get('/api/dish/getById/:id', (request, response)=>{
 const orders = require("./order.js");
 const orders_dishes = require("./orders_dishes.js");
 
-app.get('/api/order/getMaxId', function (request, response) {
-    let id = orders.getMaxId();
-    console.log(id);
-    /*.then(function(dishes){        
-        response.send(orders);
-        console.log(orders);
-    })
-    .catch(err=>response.status(400).send(err.parent.sqlMessage));*/
-    
-});
 
-app.post('/api/order/add', function (request, response) {
+app.post('/api/order/add', rutasProtegidas, function (request, response) {
     return orders.addOrder(request)
     .then(function (orders) {
         if (orders) {
             request.body.dishes.forEach(element => {
-                return dishes.getById(element.id_dish)          
+                return dishes.getById(element.id_dish)  
                 .then(function(dishes){
-                    if (dishes){
-                        return orders_dishes.addOrderDish(element)
+                    if (dishes){                                 
+                        return orders_dishes.addOrderDish(element, orders.toJSON().id)
                         .then(function(orders_dishes){
-                            response.send(orders_dishes);
-                        })
-                        .catch(err=>response.status(400).send(err.parent.sqlMessage));                         
+                            response.send(orders);
+                        })                                                          
                     }else{
                         response.status(400).send('No existe un plato con ese id' + element.id_dish);
                     }
@@ -262,49 +278,102 @@ app.post('/api/order/add', function (request, response) {
         
 });
 
-app.put('/api/order/update/:id', function (request, response) {
-    return orders.updateOrder(request)
-    .then(function (orders) {
-        if (orders) {
-            response.status(200).send('La orden ' + request.params.id + ' ha sido actualizada con éxito')
-        } else {
-            response.status(400).send('Error in update new record');
+app.put('/api/order/update', rutasProtegidas, function (request, response) {
+    return users.getByNickName(request.body.requestedBy)
+    .then(function (users) {
+        if (users){
+            if (users.toJSON().role==1){
+                return orders.updateOrder(request)
+                .then(function (orders) {
+                    if (orders) {
+                        response.status(200).send('La orden ' + request.body.id + ' ha sido actualizada con éxito')
+                    } else {
+                        response.status(400).send('Error in update new record');
+                    }
+                })   
+                .catch(err=>response.status(400).send(err.parent.sqlMessage));  
+            }else{
+                response.status(400).send('El usuario ' + request.body.requestedBy + ' no posee rol de administrador');
+            }
+        }else{
+            response.status(400).send('No existe el usuario ' + request.body.requestedBy);
         }
-    });    
+    });       
+        
 });
-app.delete('/api/order/delete/:id', function (request, response) {
-    return orders.deleteOrder(request)
-    .then(function (orders) {
-        if (orders) {
-            response.status(200).send('La orden ' + request.params.id + ' ha sido eliminada con éxito')
-        } else {
-            response.status(400).send('Error in delete new record');
+app.get('/api/order/getMaxId', rutasProtegidas, (request, response) => {
+    return users.getByNickName(request.body.requestedBy)
+    .then(function (users) {
+        if (users){
+            if (users.toJSON().role==1){
+                const orderList = orders.getAllOrders();
+                orderList.forEach(order => {
+                    if( order.id > max ) {
+                      max = order.id;
+                    }
+                });
+                console.log(max);
+            }else{
+                response.status(400).send('El usuario ' + request.body.requestedBy + ' no posee rol de administrador');
+            }
+        }else{
+            response.status(400).send('No existe el usuario ' + request.body.requestedBy);
         }
-    });    
+    });       
+        
 });
-app.get('/api/dishes/getAll', (request, response) => {
-    return dishes.getAllDishes()
-    .then(function (dishes) {
-        if (dishes) {
-            response.send(dishes);
-            
-        } else {
-            response.status(400).send('Error');
+app.get('/api/order/getAll', rutasProtegidas, (request, response) => {
+    return users.getByNickName(request.body.requestedBy)
+    .then(function (users) {
+        if (users){
+            if (users.toJSON().role==1){
+                return orders.getAllOrders()
+                .then(function (orders) {
+                    if (orders) {
+                        response.send(orders);
+                        
+                    } else {
+                        response.status(400).send('Error');
+                    }
+                })   
+                .catch(err=>response.status(400).send(err.parent.sqlMessage));  
+            }else{
+                response.status(400).send('El usuario ' + request.body.requestedBy + ' no posee rol de administrador');
+            }
+        }else{
+            response.status(400).send('No existe el usuario ' + request.body.requestedBy);
         }
-    });
-})
-app.get('/api/dish/getById/:id', (request, response)=>{
-    return dishes.getById(request)
-    .then(function (dishes) {
-        if (dishes) {
-            response.send(dishes);
-        } else {
-            response.status(400).send('Error');
+    });       
+        
+});
+app.get('/api/order/getById/:id', rutasProtegidas, (request, response)=>{
+    return users.getByNickName(request.body.requestedBy)
+    .then(function (users) {
+        if (users){
+            if (users.toJSON().role==1){
+                return orders.getById(request)
+                .then(function (orders) {
+                    if (orders) {
+                        response.send(orders);
+                    } else {
+                        response.status(400).send('Error');
+                    }
+                })   
+                .catch(err=>response.status(400).send(err.parent.sqlMessage));  
+            }else{
+                response.status(400).send('El usuario ' + request.body.requestedBy + ' no posee rol de administrador');
+            }
+        }else{
+            response.status(400).send('No existe el usuario ' + request.body.requestedBy);
         }
-    });
-})
-
+    });       
+        
+});
 
 app.listen(port, () => {
     console.log(`API REST corriendo en http://localhost:${port}`)
 })
+
+app.get('/', function(req, res) {
+    res.send('Proyecto Delilah Resto');
+});
